requirements.txt
dj-database-url==0.5.0
django==4.0.1
mysqlclient==2.1.0
djangorestframework==3.13.1
celery==5.2.3
sqlalchemy==1.4.29